Rename the new base class back to OpTestBase
