+++ tests/didt/README.md
@@ -78,6 +78,6 @@ LM head: `pytest models/demos/falcon7b/tests/test_falcon_hang.py -k "test_reprod
 
 # Adding another suspected repro test
 
-`tests/didt/op_test_base_multi.py` defines a base class for all tests that encapsulates common behavior - how we run iterations, deallocate, check determinism, sync, etc.  To add a new test, create a new file under the same directory, and then either:
+`tests/didt/op_test_base.py` defines a base class for all tests that encapsulates common behavior - how we run iterations, deallocate, check determinism, sync, etc.  To add a new test, create a new file under the same directory, and then either:
 - instantiate object of the base class in case you don't need to change any behavior, just populate dimensions, configs etc (example in `test_ff1_matmul.py`)
 - extend the base class to override any behavior that needs to be changed (for now we allow to change the way we generate activations & weights, and setting the seed), and then instantiate object of the new class
