+++ tests/didt/test_duty_cycle.py
@@ -7,7 +7,7 @@ import pytest
 import torch
 import random
 
-from tests.didt.op_test_base_multi import OpTestBaseMulti, OpParameter, get_mesh_grid_size
+from tests.didt.op_test_base import OpTestBase, OpParameter, get_mesh_grid_size
 import ttnn
 from models.common.utility_functions import skip_for_blackhole, is_blackhole, skip_for_wormhole_b0
 
@@ -16,7 +16,7 @@ MESH_X = NUM_DEVICES if NUM_DEVICES <= 8 else 8
 MESH_Y = 1 if NUM_DEVICES <= 8 else int(NUM_DEVICES / MESH_X)
 
 
-class DutyCycleTest(OpTestBaseMulti):
+class DutyCycleTest(OpTestBase):
     def __init__(self, *args, non_mm_loops=5, wl_loops=100, **kwargs):
         super().__init__(*args, **kwargs)
         self.non_mm_loops = (non_mm_loops,)
